--- a/Cargo.toml
+++ b/Cargo.toml
@@ -13,6 +13,7 @@ exclude = ["tests/*"]
 [dependencies]
 byteorder = "1.0"
 rayon = { version = "1.0", optional = true }
+packed_simd = "0.3"
 
 [dev-dependencies]
 png = "0.16"