--- a/src/idct.rs
+++ b/src/idct.rs
@@ -4,7 +4,11 @@
 use crate::parser::Dimensions;
 use std::num::Wrapping;
 
-use packed_simd::{i32x8, i16x8, u16x8, u8x8, FromCast};
+use packed_simd::{
+    i32x8, i16x8, u16x8, u8x8,
+    i32x4, i16x4, u16x4, u8x4,
+    FromCast,
+};
 
 pub(crate) fn choose_idct_size(full_size: Dimensions, requested_size: Dimensions) -> usize {
     fn scaled(len: u16, scale: usize) -> u16 { ((len as u32 * scale as u32 - 1) / 8 + 1) as u16 }
@@ -47,6 +51,29 @@ pub(crate) fn dequantize_and_idct_block(scale: usize, coefficients: &[i16], quan
     }
 }
 
+
+/// Transpose an 8x8 matrix represented by SIMD vectors
+/// This has to be a macro not to depend on const generics
+macro_rules! simd_transpose {
+    ($s:expr) => {
+        for i in 0..$s.len() {
+            for j in i + 1..$s.len() {
+                let tmp = $s[i].extract(j);
+                $s[i] = $s[i].replace(j, $s[j].extract(i));
+                $s[j] = $s[j].replace(i, tmp);
+            }
+        }
+    }
+}
+
+/// take a -128..127 value and clamp it and convert to 0..255
+macro_rules! stbi_clamp_simd {
+    ($source:tt, $target:tt, $x:expr) => (
+        $target::from_cast(
+            $x.max($source::splat(0))
+              .min($source::splat(255))))
+}
+
 // This is based on stb_image's 'stbi__idct_block'.
 fn dequantize_and_idct_block_8x8(coefficients: &[i16], quantization_table: &[u16; 64], output_linestride: usize, output: &mut [u8]) {
     debug_assert_eq!(coefficients.len(), 64);
@@ -85,7 +112,7 @@ fn dequantize_and_idct_block_8x8(coefficients: &[i16], quantization_table: &[u16
     s[7] = (x[0] - t[3]) >> 10;
 
     // columns
-    simd_transpose(&mut s);
+    simd_transpose!(s);
 
     // constants scaled things up by 1<<12, plus we had 1<<2 from first
     // loop, plus horizontal and vertical each scale by sqrt(8) so together
@@ -97,14 +124,14 @@ fn dequantize_and_idct_block_8x8(coefficients: &[i16], quantization_table: &[u16
     let t = idct_1d_t(&s);
 
     let results = [
-        stbi_clamp_simd((x[0] + t[3]) >> 17),
-        stbi_clamp_simd((x[1] + t[2]) >> 17),
-        stbi_clamp_simd((x[2] + t[1]) >> 17),
-        stbi_clamp_simd((x[3] + t[0]) >> 17),
-        stbi_clamp_simd((x[3] - t[0]) >> 17),
-        stbi_clamp_simd((x[2] - t[1]) >> 17),
-        stbi_clamp_simd((x[1] - t[2]) >> 17),
-        stbi_clamp_simd((x[0] - t[3]) >> 17),
+        stbi_clamp_simd!(i32x8,u8x8, (x[0] + t[3]) >> 17),
+        stbi_clamp_simd!(i32x8,u8x8, (x[1] + t[2]) >> 17),
+        stbi_clamp_simd!(i32x8,u8x8, (x[2] + t[1]) >> 17),
+        stbi_clamp_simd!(i32x8,u8x8, (x[3] + t[0]) >> 17),
+        stbi_clamp_simd!(i32x8,u8x8, (x[3] - t[0]) >> 17),
+        stbi_clamp_simd!(i32x8,u8x8, (x[2] - t[1]) >> 17),
+        stbi_clamp_simd!(i32x8,u8x8, (x[1] - t[2]) >> 17),
+        stbi_clamp_simd!(i32x8,u8x8, (x[0] - t[3]) >> 17),
     ];
 
     for i in 0..8 {
@@ -118,9 +145,9 @@ fn dequantize_and_idct_block_8x8(coefficients: &[i16], quantization_table: &[u16
 fn idct_1d_x(s: &[i32x8; 8], correction: i32) -> [i32x8; 4] {
     let p2 = s[2];
     let p3 = s[6];
-    let p1 = (p2 + p3) * stbi_f2f_simd(0.5411961);
-    let t2 = p1 + p3 * stbi_f2f_simd(-1.847759065);
-    let t3 = p1 + p2 * stbi_f2f_simd(0.765366865);
+    let p1 = (p2 + p3) * i32x8::splat(stbi_f2f(0.5411961));
+    let t2 = p1 + p3 * i32x8::splat(stbi_f2f(-1.847759065));
+    let t3 = p1 + p2 * i32x8::splat(stbi_f2f(0.765366865));
     let p2 = s[0];
     let p3 = s[4];
     let t0 = stbi_fsh_simd(p2 + p3);
@@ -145,86 +172,79 @@ fn idct_1d_t(s: &[i32x8; 8]) -> [i32x8; 4] {
     let p4 = t1 + t3;
     let p1 = t0 + t3;
     let p2 = t1 + t2;
-    let p5 = (p3 + p4) * stbi_f2f_simd(1.175875602);
-    let t0 = t0 * stbi_f2f_simd(0.298631336);
-    let t1 = t1 * stbi_f2f_simd(2.053119869);
-    let t2 = t2 * stbi_f2f_simd(3.072711026);
-    let t3 = t3 * stbi_f2f_simd(1.501321110);
-    let p1 = p5 + (p1 * stbi_f2f_simd(-0.899976223));
-    let p2 = p5 + (p2 * stbi_f2f_simd(-2.562915447));
-    let p3 = p3 * stbi_f2f_simd(-1.961570560);
-    let p4 = p4 * stbi_f2f_simd(-0.390180644);
+    let p5 = (p3 + p4) * i32x8::splat(stbi_f2f(1.175875602));
+    let t0 = t0 * i32x8::splat(stbi_f2f(0.298631336));
+    let t1 = t1 * i32x8::splat(stbi_f2f(2.053119869));
+    let t2 = t2 * i32x8::splat(stbi_f2f(3.072711026));
+    let t3 = t3 * i32x8::splat(stbi_f2f(1.501321110));
+    let p1 = p5 + (p1 * i32x8::splat(stbi_f2f(-0.899976223)));
+    let p2 = p5 + (p2 * i32x8::splat(stbi_f2f(-2.562915447)));
+    let p3 = p3 * i32x8::splat(stbi_f2f(-1.961570560));
+    let p4 = p4 * i32x8::splat(stbi_f2f(-0.390180644));
     [t0 + p1 + p3, t1 + p2 + p4, t2 + p2 + p3, t3 + p1 + p4, ]
-}
-
-/// Transpose an 8x8 matrix represented by SIMD vectors
-#[inline(always)]
-fn simd_transpose(s: &mut [i32x8; 8]) {
-    for i in 0..8 {
-        for j in i + 1..8 {
-            let tmp = s[i].extract(j);
-            s[i] = s[i].replace(j, s[j].extract(i));
-            s[j] = s[j].replace(i, tmp);
-        }
-    }
 }
 
 // 4x4 and 2x2 IDCT based on Rakesh Dugad and Narendra Ahuja: "A Fast Scheme for Image Size Change in the Compressed Domain" (2001).
 // http://sylvana.net/jpegcrop/jidctred/
 fn dequantize_and_idct_block_4x4(coefficients: &[i16], quantization_table: &[u16; 64], output_linestride: usize, output: &mut [u8]) {
     debug_assert_eq!(coefficients.len(), 64);
-    let mut temp = [Wrapping(0i32); 4 * 4];
 
-    const CONST_BITS: usize = 12;
-    const PASS1_BITS: usize = 2;
-    const FINAL_BITS: usize = CONST_BITS + PASS1_BITS + 3;
+    const CONST_BITS: u32 = 12;
+    const PASS1_BITS: u32 = 2;
+    const FINAL_BITS: u32 = CONST_BITS + PASS1_BITS + 3;
+
+    let mut s: [i32x4; 4] = [
+        i32x4::from(i16x4::from_slice_unaligned(&coefficients[0..0 + 8])) *
+            i32x4::from(u16x4::from_slice_unaligned(&quantization_table[0..0 + 4])),
+        i32x4::from(i16x4::from_slice_unaligned(&coefficients[8..8 + 4])) *
+            i32x4::from(u16x4::from_slice_unaligned(&quantization_table[8..8 + 4])),
+        i32x4::from(i16x4::from_slice_unaligned(&coefficients[16..16 + 4])) *
+            i32x4::from(u16x4::from_slice_unaligned(&quantization_table[16..16 + 4])),
+        i32x4::from(i16x4::from_slice_unaligned(&coefficients[24..24 + 4])) *
+            i32x4::from(u16x4::from_slice_unaligned(&quantization_table[24..24 + 4])),
+    ];
+
+    let x0 = (s[0] + s[2]) << PASS1_BITS;
+    let x2 = (s[0] - s[2]) << PASS1_BITS;
+    let p1 = (s[1] + s[3]) * i32x4::splat(stbi_f2f(0.541196100));
+    let t0 = (p1
+        + s[3] * i32x4::splat(stbi_f2f(-1.847759065))
+        + i32x4::splat(512)
+    ) >> (CONST_BITS - PASS1_BITS);
+    let t2 = (p1
+        + s[1] * i32x4::splat(stbi_f2f(0.765366865))
+        + i32x4::splat(512)
+    ) >> (CONST_BITS - PASS1_BITS);
+
+    s[0] = x0 + t2;
+    s[1] = x2 + t0;
+    s[2] = x2 - t0;
+    s[3] = x0 - t2;
 
     // columns
-    for i in 0..4 {
-        let s0 = Wrapping(coefficients[i + 8 * 0] as i32 * quantization_table[i + 8 * 0] as i32);
-        let s1 = Wrapping(coefficients[i + 8 * 1] as i32 * quantization_table[i + 8 * 1] as i32);
-        let s2 = Wrapping(coefficients[i + 8 * 2] as i32 * quantization_table[i + 8 * 2] as i32);
-        let s3 = Wrapping(coefficients[i + 8 * 3] as i32 * quantization_table[i + 8 * 3] as i32);
+    simd_transpose!(s);
 
-        let x0 = (s0 + s2) << PASS1_BITS;
-        let x2 = (s0 - s2) << PASS1_BITS;
+    let x0 = (s[0] + s[2]) << CONST_BITS;
+    let x2 = (s[0] - s[2]) << CONST_BITS;
 
-        let p1 = (s1 + s3) * stbi_f2f(0.541196100);
-        let t0 = (p1 + s3 * stbi_f2f(-1.847759065) + Wrapping(512)) >> (CONST_BITS - PASS1_BITS);
-        let t2 = (p1 + s1 * stbi_f2f(0.765366865) + Wrapping(512)) >> (CONST_BITS - PASS1_BITS);
+    let p1 = (s[1] + s[3]) * i32x4::splat(stbi_f2f(0.541196100));
+    let t0 = p1 + s[3] * i32x4::splat(stbi_f2f(-1.847759065));
+    let t2 = p1 + s[1] * i32x4::splat(stbi_f2f(0.765366865));
 
-        temp[i + 4 * 0] = x0 + t2;
-        temp[i + 4 * 3] = x0 - t2;
-        temp[i + 4 * 1] = x2 + t0;
-        temp[i + 4 * 2] = x2 - t0;
-    }
+    let x0 = x0 + i32x4::splat(1 << (FINAL_BITS - 1)) + i32x4::splat(128 << FINAL_BITS);
+    let x2 = x2 + i32x4::splat(1 << (FINAL_BITS - 1)) + i32x4::splat(128 << FINAL_BITS);
 
-    for i in 0 .. 4 {
-        let s0 = temp[i * 4 + 0];
-        let s1 = temp[i * 4 + 1];
-        let s2 = temp[i * 4 + 2];
-        let s3 = temp[i * 4 + 3];
+    let results = [
+        stbi_clamp_simd!(i32x4,u8x4, (x0 + t2) >> FINAL_BITS),
+        stbi_clamp_simd!(i32x4,u8x4, (x2 + t0) >> FINAL_BITS),
+        stbi_clamp_simd!(i32x4,u8x4, (x2 - t0) >> FINAL_BITS),
+        stbi_clamp_simd!(i32x4,u8x4, (x0 - t2) >> FINAL_BITS),
+    ];
 
-        let x0 = (s0 + s2) << CONST_BITS;
-        let x2 = (s0 - s2) << CONST_BITS;
-
-        let p1 = (s1 + s3) * stbi_f2f(0.541196100);
-        let t0 = p1 + s3 * stbi_f2f(-1.847759065);
-        let t2 = p1 + s1 * stbi_f2f(0.765366865);
-
-        // constants scaled things up by 1<<12, plus we had 1<<2 from first
-        // loop, plus horizontal and vertical each scale by sqrt(8) so together
-        // we've got an extra 1<<3, so 1<<17 total we need to remove.
-        // so we want to round that, which means adding 0.5 * 1<<17,
-        // aka 65536. Also, we'll end up with -128 to 127 that we want
-        // to encode as 0..255 by adding 128, so we'll add that before the shift
-        let x0 = x0 + Wrapping(1 << (FINAL_BITS - 1)) + Wrapping(128 << FINAL_BITS);
-        let x2 = x2 + Wrapping(1 << (FINAL_BITS - 1)) + Wrapping(128 << FINAL_BITS);
-
-        output[i * output_linestride + 0] = stbi_clamp((x0 + t2) >> FINAL_BITS);
-        output[i * output_linestride + 3] = stbi_clamp((x0 - t2) >> FINAL_BITS);
-        output[i * output_linestride + 1] = stbi_clamp((x2 + t0) >> FINAL_BITS);
-        output[i * output_linestride + 2] = stbi_clamp((x2 - t0) >> FINAL_BITS);
+    for i in 0..results.len() {
+        for j in 0..results.len() {
+            output[i * output_linestride + j] = results[j].extract(i);
+        }
     }
 }
 
@@ -272,18 +292,8 @@ fn stbi_clamp(x: Wrapping<i32>) -> u8
     x.0.max(0).min(255) as u8
 }
 
-fn stbi_f2f(x: f32) -> Wrapping<i32> {
-    Wrapping((x * 4096.0 + 0.5) as i32)
-}
-
-// take a -128..127 value and stbi__clamp it and convert to 0..255
-fn stbi_clamp_simd(x: i32x8) -> u8x8
-{
-    u8x8::from_cast(x.max(i32x8::splat(0)).min(i32x8::splat(255)))
-}
-
-fn stbi_f2f_simd(x: f32) -> i32x8 {
-    i32x8::splat((x * 4096.0 + 0.5) as i32)
+fn stbi_f2f(x: f32) -> i32 {
+    (x * 4096.0 + 0.5) as i32
 }
 
 fn stbi_fsh_simd(x: i32x8) -> i32x8 {