@lovasoa 2020-04-16
Final benchmark 

## Before this PR
```
$ RUSTFLAGS="-C llvm-args=--vectorize-slp -C target-cpu=native" cargo bench 'decode a 512x512 JPEG'

decode a 512x512 JPEG   time:   [2.6583 ms 2.6783 ms 2.6984 ms]      
```

## After this PR
### all optimizations, `packed_simd` feature disabled

```
$ RUSTFLAGS="-C llvm-args=--vectorize-slp -C target-cpu=native" cargo bench 'decode a 512x512 JPEG'

decode a 512x512 JPEG   time:   [3.5087 ms 3.5182 ms 3.5281 ms]                                   
```

**~ 30% worse**

### all optimizations, `packed_simd` feature enabled

```
$ RUSTFLAGS="-C llvm-args=--vectorize-slp -C target-cpu=native" cargo +nightly bench --features="packed_simd" 'decode a 512x512 JPEG'

decode a 512x512 JPEG   time:   [2.0040 ms 2.0151 ms 2.0262 ms]                                   
```

**~ 20% better**

So contrarily to what I thought, the benchmarks didn't improve, they worsened for the non-simd case.