@lovasoa 2020-06-15
Hi @HeroicKatora ! 
  Yes, there is a performance regression for the non-simd version, so this PR should probably not be merged as-is. I contacted @lilith about this PR, but she told me her company doesn't need this anymore and thus isn't ready to fund this anymore, so I lost a little bit of the motivation needed to work on this. 

  I wanted to switch from packed_simd to [simdeez](https://crates.io/crates/simdeez), but then I found an inconsistency in simdeez, for which I proposed a [PR](https://github.com/jackmott/simdeez/pull/28). It took some time, but it is now merged, so there is no more blocker. Are you interested in working on this, @HeroicKatora ?