@HeroicKatora 2020-06-15
I see, and can totally feel how that would turn off some motivation. Yeah, the total of 20% performance is signifcant. Switching to a stable crate with `1.0` would be positive as well and if it helps avoid the regression even better!

I'm looking for what project to focus on next after `apng` is done, so maybe. The alternative is that I focus on [`image-canvas`] instead, still shepherding this if you think it has the potential.